partially working and compiling
