+++ src/app/main/utility-components/track-data-utility/track-data/track-data-table/track-data-table.component.ts
@@ -5,8 +5,8 @@ import { AstroComponentsModule } from '@astrouxds/angular';
 import { SitesComponent } from '../sites/sites.component';
 import { SettingsComponent } from '../settings/settings.component';
 import { TrackFile } from 'src/app/types/data.types';
-import { TrackFilesDataUtilityService } from '../../track-files-data.service';
-import { TableService } from 'src/app/shared/table.service';
+import { TrackData, TrackFilesDataUtilityService } from '../../track-files-data.service';
+import { Column, ColumnDefs, TableService } from 'src/app/shared/table.service';
 
 @Component({
   selector: 'fds-track-data-table',
@@ -17,39 +17,14 @@ import { TableService } from 'src/app/shared/table.service';
     SitesComponent,
     SettingsComponent,
   ],
-  providers: [TableService],
   templateUrl: './track-data-table.component.html',
   styleUrls: ['./track-data-table.component.css'],
 })
 export class TrackDataTableComponent {
+  tableColumns: Column<TrackData>[]
 
-  sharedTableData$ = this.trackFilesService.get()
-  sharedTableData: any[] = []
-
-  //table variables
-  columnDefs = [
-    { header: 'File Name', field: 'name', sortable: true },
-    { header: 'Date', field: 'creationDate', sortable: true },
-    { header: 'File Size', field: 'fileSize', sortable: true },
-  ];
-
-  //subscription cleanup
-  destroyRef = inject(DestroyRef)
-  destroyed = new Subject();
-
-  constructor(public trackFilesService: TrackFilesDataUtilityService, public tableService: TableService<any>){}
-
-  ngOnInit(){
-    this.sharedTableData$.subscribe(res => {
-      this.sharedTableData = res})
-    this.getFiles()
-  }
-  getFiles(){
-    this.tableService.init({data: this.sharedTableData, columnDefs: this.columnDefs});
+  constructor(public trackFilesService: TrackFilesDataUtilityService){
+    this.tableColumns = trackFilesService.tableService.columns.filter(column => column.field !== 'id')
   }
 
-  ngOnDestroy(){
-    this.destroyed.next(true);
-    this.destroyed.complete();
-  }
 }
